DFE-837 - restructured services and mocks
