+++ src/explore-education-statistics-admin/src/components/DashboardRelease.tsx
@@ -5,7 +5,10 @@ import Link from '@admin/components/Link';
 import { format } from 'date-fns';
 import Details from '@common/components/Details';
 import { LoginContext } from '@admin/components/Login';
-import { ApprovalStatus, Release } from '@admin/services/types/types';
+import {
+  ApprovalStatus,
+  Release,
+} from '@admin/services/api/common/types/types';
 import { setupRoute } from '@admin/routes/releaseRoutes';
 
 const getLiveLatestLabel = (isLive: boolean, isLatest: boolean) => {
