+++ src/explore-education-statistics-admin/src/services/api/dashboard/mock/axios-mock.ts
@@ -1,11 +1,14 @@
-import DummyPublicationsData from '@admin/pages/DummyPublicationsData';
-import { PrototypeLoginService } from '@admin/services/PrototypeLoginService';
 import { AxiosInstance } from 'axios';
+import { PrototypeLoginService } from '../../../PrototypeLoginService';
 
 export default {
   createMockContentApiAxiosInstance: async (axiosInstance: AxiosInstance) => {
     const MockAdaptor = (await import(
       /* webpackChunkName: "axios-mock-adapter" */ 'axios-mock-adapter'
+    )).default;
+
+    const mockData = (await import(
+      /* webpackChunkName: "mock-data" */ './mock-data'
     )).default;
 
     const mock = new MockAdaptor(axiosInstance);
@@ -15,7 +18,7 @@ export default {
       .onGet('/Theme', {
         params: { userId: PrototypeLoginService.getUserList()[0].id },
       })
-      .reply(200, DummyPublicationsData.themesAndTopics);
+      .reply(200, mockData.themesAndTopics);
 
     return axiosInstance;
   },
