+++ src/explore-education-statistics-admin/src/services/api/index.ts
@@ -1,8 +1,7 @@
-import { ThemeAndTopics } from '@admin/pages/DummyPublicationsData';
 import Client from '@common/services/api/Client';
 import axios from 'axios';
 import { commaSeparated } from '@common/services/util/paramSerializers';
-import mocks from './mock/axios-mock';
+import mocks from './dashboard/mock/axios-mock';
 
 const createContentApiAxios = async () => {
   const baseUrl = process.env.CONTENT_API_BASE_URL;
@@ -19,10 +18,6 @@ const createContentApiAxios = async () => {
 
 export default {
   contentApi: createContentApiAxios().then(axiosInstance => {
-    const client = new Client(axiosInstance);
-    return {
-      getThemesAndTopics: (userId: string) =>
-        client.get<ThemeAndTopics[]>('/Theme', { params: { userId } }),
-    };
+    return new Client(axiosInstance);
   }),
 };
